[GOBBLIN-1465] Refactor the GobblinOrcWriter to support using a different OrcValueWriter
